cmd/genblog: rename from gen
Make room for `cmd/gendoc`, a static docs generator.
Ignore one copy of the binaries.
Generate a better README co-located in blog template.
