+++ cmd/genblog/.gitignore
@@ -1,0 +1,4 @@
+blog.tar.gz
+blog/README.md
+blog/bin/linux
+blog/bin/mac
