+++ cmd/genblog/blog/.gitignore
@@ -1,0 +1,2 @@
+public/*
+!public/images
