+++ cmd/genblog/blog/config.json
@@ -1,0 +1,4 @@
+{
+  "name": "Blog",
+  "url": "https://blog.example.com"
+}
