+++ cmd/genblog/blog/genblog
@@ -1,0 +1,19 @@
+#!/usr/bin/env bash
+
+# Shim to architecture-specific binary
+# https://github.com/statusok/statusok/tree/master/genblog
+
+os="$(uname -s)"
+arch="$(uname -m)"
+
+if [[ "$os" == "Linux" && "$arch" == "x86_64" ]]; then
+  bin/linux "$@"
+elif [[ "$os" == "Darwin" && "$arch" == "x86_64" ]]; then
+  bin/mac "$@"
+else
+  echo "$os/$arch is not supported. If you have Go installed, run:"
+  echo ""
+  echo "  go get github.com/statusok/statusok/genblog"
+  echo ""
+  echo "Otherwise, get help at https://github.com/statusok/statusok/issues"
+fi
