--- gen/blog/.gitignore
@@ -1,2 +1,0 @@
-public/*
-!public/images
