--- gen/blog/README.md
@@ -1,9 +1,0 @@
-# Blog
-
-A static blog.
-
-## Workflow
-
-See [documentation][docs].
-
-[docs]: https://github.com/statusok/statusok/tree/master/gen
