--- gen/blog/gen.json
@@ -1,4 +1,0 @@
-{
-  "name": "Blog",
-  "url": "https://blog.example.com"
-}
